+++ src/test/java/tokyo/northside/io/FileUtils2Test.java
@@ -9,14 +9,14 @@ import java.io.File;
 /**
  * Created by Hiroshi Miura on 16/04/09.
  */
-public class FileUtilsTest {
+public class FileUtils2Test {
 
     @Test (groups="file")
     public void testContentEquals_range() throws Exception {
         System.out.println("contentEquals_range");
         File firstFile = new File(this.getClass().getResource("/test_util.txt").getFile());
         File secondFile = new File(this.getClass().getResource("/test_util2.txt").getFile());
-        assertTrue(FileUtils.contentEquals(firstFile, secondFile, 10, 64));
+        assertTrue(FileUtils2.contentEquals(firstFile, secondFile, 10, 64));
     }
 
     @Test (groups="file")
@@ -24,6 +24,6 @@ public class FileUtilsTest {
         System.out.println("contentEqualsEquals_range_false");
         File firstFile = new File(this.getClass().getResource("/test_util.txt").getFile());
         File secondFile = new File(this.getClass().getResource("/test_util2.txt").getFile());
-        assertFalse(FileUtils.contentEquals(firstFile, secondFile, 0, 64));
+        assertFalse(FileUtils2.contentEquals(firstFile, secondFile, 0, 64));
     }
 }